feat: move class wrapper to another file and add prefix
