+++ src/routes/index.ts
@@ -1,4 +1,4 @@
-import Koa, { Context } from 'koa';
+import Koa from 'koa';
 import Router from 'koa-router';
 import requireDir from 'require-dir';
 import urljoin from 'url-join';
@@ -8,7 +8,7 @@ interface RouteConfig {
   prefix: string;
 }
 
-interface App {
+interface Context {
   app: Koa;
 }
 
@@ -18,7 +18,7 @@ interface Dir {
 
 export type TRegisterRoute = () => Router;
 
-export default function registerRoutes(ctx: App) {
+export default function registerRoutes(ctx: Context) {
   const { app } = ctx;
   app.use(
     error({
@@ -49,7 +49,7 @@ export default function registerRoutes(ctx: App) {
   app.use(router.routes());
 }
 
-function registerDir(dir: Dir, ctx: App, config: RouteConfig) {
+function registerDir(dir: Dir, ctx: Context, config: RouteConfig) {
   const { app } = ctx;
   Object.keys(dir).forEach((key) => {
     let routePrefix = urljoin(config.prefix, key);
@@ -61,47 +61,13 @@ function registerDir(dir: Dir, ctx: App, config: RouteConfig) {
       registerDir(required, ctx, {
         prefix: routePrefix,
       });
-      return;
     } else if (typeof required.default === 'function') {
       const registerRoute = required.default as TRegisterRoute;
       const router = registerRoute();
       console.info('route added: ', routePrefix);
       app.use(router.routes());
-      return;
+    } else {
+      throw new Error(`route file is incorrect: ${routePrefix}`);
     }
-    throw new Error(`route file is incorrect: ${routePrefix}`);
   });
-}
-
-type PathParams<Path extends string> = Path extends `:${infer Param}/${infer Rest}`
-  ? Param | PathParams<Rest>
-  : Path extends `:${infer Param}`
-  ? Param
-  : Path extends `${infer _Prefix}:${infer Rest}`
-  ? PathParams<`:${Rest}`>
-  : never;
-
-type PathArgs<Path extends string> = {
-  [K in PathParams<Path>]: string;
-};
-export class ParamedRouters<T, K> extends Router {
-  paramedGet<P extends string>(path: P, handler: (ctx: Context & { params: PathArgs<P> }) => Promise<void>): Router {
-    return this.get(path, handler);
-  }
-
-  paramedPost<P extends string>(path: P, handler: (ctx: Context & { params: PathArgs<P> }) => Promise<void>): Router {
-    return this.post(path, handler);
-  }
-
-  paramedDel<P extends string>(path: P, handler: (ctx: Context & { params: PathArgs<P> }) => Promise<void>): Router {
-    return this.del(path, handler);
-  }
-
-  paramedPut<P extends string>(path: P, handler: (ctx: Context & { params: PathArgs<P> }) => Promise<void>): Router {
-    return this.put(path, handler);
-  }
-
-  paramedAll<P extends string>(path: P, handler: (ctx: Context & { params: PathArgs<P> }) => Promise<void>): Router {
-    return this.all(path, handler);
-  }
 }
